+++ noodles-sam/src/record/data/field/value/subtype.rs
@@ -68,14 +68,20 @@ impl FromStr for Subtype {
 
 impl From<Subtype> for char {
     fn from(subtype: Subtype) -> Self {
+        char::from(u8::from(subtype))
+    }
+}
+
+impl From<Subtype> for u8 {
+    fn from(subtype: Subtype) -> Self {
         match subtype {
-            Subtype::Int8 => 'c',
-            Subtype::UInt8 => 'C',
-            Subtype::Int16 => 's',
-            Subtype::UInt16 => 'S',
-            Subtype::Int32 => 'i',
-            Subtype::UInt32 => 'I',
-            Subtype::Float => 'f',
+            Subtype::Int8 => b'c',
+            Subtype::UInt8 => b'C',
+            Subtype::Int16 => b's',
+            Subtype::UInt16 => b'S',
+            Subtype::Int32 => b'i',
+            Subtype::UInt32 => b'I',
+            Subtype::Float => b'f',
         }
     }
 }
@@ -122,5 +128,16 @@ mod tests {
         assert_eq!(char::from(Subtype::Int32), 'i');
         assert_eq!(char::from(Subtype::UInt32), 'I');
         assert_eq!(char::from(Subtype::Float), 'f');
+    }
+
+    #[test]
+    fn test_from_subtype_for_u8() {
+        assert_eq!(u8::from(Subtype::Int8), b'c');
+        assert_eq!(u8::from(Subtype::UInt8), b'C');
+        assert_eq!(u8::from(Subtype::Int16), b's');
+        assert_eq!(u8::from(Subtype::UInt16), b'S');
+        assert_eq!(u8::from(Subtype::Int32), b'i');
+        assert_eq!(u8::from(Subtype::UInt32), b'I');
+        assert_eq!(u8::from(Subtype::Float), b'f');
     }
 }